+++ app/javascript/vue/tasks/digitize/store/actions/resetWithDefault.js
@@ -1,6 +1,8 @@
 import ActionNames from './actionNames'
 
 export default ({ dispatch, state }) => {
+  const { locked } = state.settings
+
   dispatch(ActionNames.NewCollectionEvent)
   dispatch(ActionNames.NewCollectionObject)
   dispatch(ActionNames.NewTypeMaterial)
@@ -18,5 +20,7 @@ export default ({ dispatch, state }) => {
   state.materialTypes = []
   state.preparation_type_id = undefined
   state.taxon_determinations = []
-  state.biologicalAssociations = []
+  state.biologicalAssociations = locked.biologicalAssociations
+    ? state.biologicalAssociations.map(item => ({...item, id: undefined, global_id: undefined }))
+    : []
 }