+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ### Added
 
 - Content smart selector
+- Biological association list lock button in comprehensive task
 
 ### Changed